fix: move config selectors
