+++ src/App.tsx
@@ -13,7 +13,13 @@ import semver from 'semver';
 
 import {TelemetryDocumentationUrl} from '@constants/tooltips';
 
-import {setCreateProject, setDeleteProject, setLoadingProject, setOpenProject} from '@redux/appConfig';
+import {
+  activeProjectSelector,
+  setCreateProject,
+  setDeleteProject,
+  setLoadingProject,
+  setOpenProject,
+} from '@redux/appConfig';
 import {toggleForm} from '@redux/forms';
 import {setIsGitInstalled} from '@redux/git';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
@@ -44,7 +50,6 @@ import {NewVersionCode, Project} from '@shared/models/config';
 import {Size} from '@shared/models/window';
 import electronStore from '@shared/utils/electronStore';
 import {setMainProcessEnv} from '@shared/utils/env';
-import {activeProjectSelector} from '@shared/utils/selectors';
 
 import * as S from './App.styled';
 import AppContext from './AppContext';
