+++ src/components/organisms/HotKeysHandler/HotKeysHandler.tsx
@@ -4,6 +4,13 @@ import {useStore} from 'react-redux';
 
 import {makeApplyKustomizationText, makeApplyResourceText} from '@constants/makeApplyText';
 
+import {
+  currentConfigSelector,
+  kubeConfigContextColorSelector,
+  kubeConfigContextSelector,
+  kubeConfigPathSelector,
+  kubeConfigPathValidSelector,
+} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {openResourceDiffModal, resetResourceFilter} from '@redux/reducers/main';
 import {
@@ -16,11 +23,8 @@ import {
   toggleRightMenu,
 } from '@redux/reducers/ui';
 import {
-  currentConfigSelector,
   isInClusterModeSelector,
   isInPreviewModeSelectorNew,
-  kubeConfigContextColorSelector,
-  kubeConfigPathSelector,
   rootFilePathSelector,
   selectedFilePathSelector,
 } from '@redux/selectors';
@@ -45,7 +49,6 @@ import {useSelectorWithRef} from '@utils/hooks';
 import {hotkeys} from '@shared/constants/hotkeys';
 import {RootState} from '@shared/models/rootState';
 import {selectFromHistory} from '@shared/utils/selectionHistory';
-import {kubeConfigContextSelector, kubeConfigPathValidSelector} from '@shared/utils/selectors';
 
 const HotKeysHandler = () => {
   const store = useStore();
