+++ src/components/organisms/LocalResourceDiffModal/LocalResourceDiffModal.tsx
@@ -17,16 +17,17 @@ import {
   makeReplaceResourceText,
 } from '@constants/makeApplyText';
 
-import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {setAlert} from '@redux/reducers/alert';
-import {closeResourceDiffModal, openResourceDiffModal} from '@redux/reducers/main';
 import {
   currentClusterAccessSelector,
   currentConfigSelector,
-  isInClusterModeSelector,
   kubeConfigContextColorSelector,
+  kubeConfigContextSelector,
   kubeConfigPathSelector,
-} from '@redux/selectors';
+} from '@redux/appConfig';
+import {useAppDispatch, useAppSelector} from '@redux/hooks';
+import {setAlert} from '@redux/reducers/alert';
+import {closeResourceDiffModal, openResourceDiffModal} from '@redux/reducers/main';
+import {isInClusterModeSelector} from '@redux/selectors';
 import {useActiveResourceMap} from '@redux/selectors/resourceMapSelectors';
 import {isKustomizationResource} from '@redux/services/kustomize';
 import {applyResource} from '@redux/thunks/applyResource';
@@ -43,7 +44,6 @@ import {getResourceKindHandler} from '@src/kindhandlers';
 import {Icon} from '@monokle/components';
 import {AlertEnum, AlertType} from '@shared/models/alert';
 import {createKubeClient, hasAccessToResourceKind} from '@shared/utils/kubeclient';
-import {kubeConfigContextSelector} from '@shared/utils/selectors';
 
 import * as S from './styled';
