+++ src/components/organisms_new/PaneManager/PaneManager.tsx
@@ -2,6 +2,7 @@ import React, {useCallback, useMemo} from 'react';
 
 import {Skeleton} from 'antd';
 
+import {activeProjectSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setPaneConfiguration, toggleLeftMenu} from '@redux/reducers/ui';
 import {isInClusterModeSelector} from '@redux/selectors';
@@ -12,7 +13,6 @@ import {EmptyDashboard} from '@organisms/Dashboard/EmptyDashboard';
 import {useMainPaneDimensions} from '@utils/hooks';
 
 import {ResizableColumnsPanel, ResizableRowsPanel} from '@monokle/components';
-import {activeProjectSelector} from '@shared/utils/selectors';
 
 import ProblemPane from '../ProblemPane';
 import StartPage from '../StartPage';
