+++ src/components/organisms_new/SettingsPane/CurrentProjectSettings/CurrentProjectSettings.tsx
@@ -1,9 +1,12 @@
-import {changeCurrentProjectName, updateProjectConfig} from '@redux/appConfig';
+import {
+  activeProjectSelector,
+  changeCurrentProjectName,
+  currentConfigSelector,
+  updateProjectConfig,
+} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {currentConfigSelector} from '@redux/selectors';
 
 import {Project, ProjectConfig} from '@shared/models/config';
-import {activeProjectSelector} from '@shared/utils/selectors';
 
 import {Settings} from '../Settings/Settings';
