+++ src/navsections/ImagesSectionBlueprint/ImagesSectionNameDisplay.tsx
@@ -1,9 +1,10 @@
 import {useMemo, useState} from 'react';
 import {useDebounce} from 'react-use';
 
+import {kubeConfigContextColorSelector} from '@redux/appConfig';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
 import {setImagesSearchedValue} from '@redux/reducers/main';
-import {isInClusterModeSelector, isInPreviewModeSelectorNew, kubeConfigContextColorSelector} from '@redux/selectors';
+import {isInClusterModeSelector, isInPreviewModeSelectorNew} from '@redux/selectors';
 
 import {SectionInstance} from '@shared/models/navigator';
