+++ src/redux/appConfig/index.ts
@@ -1,1 +1,2 @@
 export * from './appConfig.slice';
+export * from './appConfig.selectors';
