+++ src/redux/thunks/multiplePathsChanged.ts
@@ -2,8 +2,8 @@ import {createAsyncThunk, createNextState} from '@reduxjs/toolkit';
 
 import micromatch from 'micromatch';
 
+import {currentConfigSelector} from '@redux/appConfig';
 import {setChangedFiles, setGitLoading} from '@redux/git';
-import {currentConfigSelector} from '@redux/selectors';
 import {addPath, getFileEntryForAbsolutePath, reloadFile} from '@redux/services/fileEntry';
 
 import {promiseFromIpcRenderer} from '@utils/promises';
